move reused logic to BaseAnalyzer
